requirements.txt
flask==3.0.2
torch==2.2.1
torchvision==0.17.1
pillow==10.2.0
pymongo==4.6.1
python-dotenv==1.0.1
pyjwt==2.8.0
werkzeug==3.0.1
requests==2.31.0 